Commit #6
Robot can't walk through walls anymore, I did this by making a method to track where the robot position is on the x and y axis, and how big the box is. The robot will not walk through any walls, it will report that it can't walk through any walls if a command is given that will collide will the wall.