Commit #3
I added two children classes, obstacle and goals, to the Room class. The Robot class uses the Room class, I made a test Class so I can make sure the code works with all of the different classes, so far it does. I just used the same methods of loops and running the simulation from the old procedural programming version. I made a entity class for the locations variable.