Commit #5
Added a goal, and I made sure to use the goal class more than I did with the obstacle class. I made methods in the room class using the goal class, and then added them to the test class and it all works. The game ends and prints the credits when the robot reaches the goal.