Requirement 1
I added the undo button and rewind. I used a linked list and nodes to add to a stack of commands. It took a lot longer than thought it would, but I got it working for the most part. The undo command will place the robot in its previous coordinate and direction. The rewind will rewind through all commands used stored in the stack by removing the top value. Then it will place the robot in the original coordinate and direction. In my tests it's worked almost every time.