Requirement 2
I added the batch command. I find this far easier than the rewind command to implement. I had to rework a lot of my old methods of moving the robot around. In the test method it will add all of the user input by using the split command to seperate the commands into a list. From all of my tests it seems to work everytime.